feat(plugin): 範囲外チェックと例のコードを追加

+++ kernel/add.cl
@@ -1,6 +1,6 @@
 // add: c = a + b を計算します。
 
 kernel void run(constant float *a, constant float *b, global float *c) {
-  int id = get_global_id(0);
+  size_t id = get_global_id(0);
   c[id] = a[id] + b[id];
 }
